update tryme and tryme2 new CMakeLists.txt
